+++ templates/read.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Chapter {{ chapter }}</title>
+    <style>
+        body { margin: 0; padding: 0; background: #f0f0f0; }
+        img { 
+            display: block; 
+            max-width: 100%; 
+            height: auto; 
+            margin: 20px auto; 
+            box-shadow: 0 2px 5px rgba(0,0,0,0.2);
+            border-radius: 5px;
+        }
+        .back-button {
+            position: fixed;
+            top: 20px;
+            left: 20px;
+            padding: 10px 20px;
+            background: #3498db;
+            color: white;
+            text-decoration: none;
+            border-radius: 5px;
+            z-index: 1000;
+        }
+    </style>
+</head>
+<body>
+    <a href="/" class="back-button">← Back to Chapters</a>
+    <div id="image-container"></div>
+
+    <script>
+        let chapter = {{ chapter }};
+        let start = 0;
+        const count = 5;
+        let loading = false;
+
+        function loadImages() {
+            if (loading) return;
+            loading = true;
+            
+            fetch(`/load_images?chapter=${chapter}&start=${start}&count=${count}`)
+                .then(r => r.json())
+                .then(data => {
+                    data.images.forEach(img => {
+                        const imageEl = document.createElement('img');
+                        imageEl.src = `/chapter_image/${chapter}/${img.filename}`;
+                        imageEl.loading = 'lazy';
+                        document.getElementById('image-container').appendChild(imageEl);
+                    });
+                    
+                    if (data.next_start !== null) {
+                        start = data.next_start;
+                        loading = false;
+                    }
+                });
+        }
+
+        window.addEventListener('scroll', () => {
+            if (window.innerHeight + window.scrollY >= document.body.offsetHeight - 500) {
+                loadImages();
+            }
+        });
+
+        // Initial load
+        loadImages();
+    </script>
+</body>
+</html>
